succesfully implemented user profile picture, it can be saved and retrieved from the database. MyImageDisplayer and MyImageProcessor have changes that some of its codes are commented out for now as they are moved to the HomeAppBarProvider.
