+++ lib/view/home/widgets/appbar/home_appbar.dart
@@ -1,9 +1,13 @@
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
-import 'package:wanderhuman_app/helper/firebase_services.dart';
+import 'package:provider/provider.dart';
+import 'package:wanderhuman_app/helper/personal_info_repository.dart';
 import 'package:wanderhuman_app/model/personal_info.dart';
 import 'package:wanderhuman_app/utilities/properties/dimension_adapter.dart';
-import 'package:wanderhuman_app/view/components/appbar/menu_options.dart';
+import 'package:wanderhuman_app/view-model/home_appbar_provider.dart';
+import 'package:wanderhuman_app/view/components/image_displayer.dart';
+import 'package:wanderhuman_app/view/home/widgets/appbar/menu_options.dart';
+import 'package:wanderhuman_app/view/home/widgets/home_utility_functions/profile_picture_bottom_modal_sheet.dart';
 
 class HomeAppBar extends StatefulWidget {
   const HomeAppBar({super.key});
@@ -24,17 +28,18 @@ class _HomeAppBarState extends State<HomeAppBar> {
 
   // fetches all the users from firestore through MyFirebaseServices
   Future<void> fetchUsers() async {
-    usersList = await MyFirebaseServices.getAllPersonalInfoRecords();
+    usersList = await MyPersonalInfoRepository.getAllPersonalInfoRecords();
   }
 
   // to get the current user's name and UPDATE the userName variable's state
   Future<void> fetchAndSetUsername() async {
     try {
       await fetchUsers();
-      String name = MyFirebaseServices.getSpecificUserNameOfTheLoggedInAccount(
-        personsList: usersList,
-        userIDToLookFor: FirebaseAuth.instance.currentUser!.uid,
-      );
+      String name =
+          MyPersonalInfoRepository.getSpecificUserNameOfTheLoggedInAccount(
+            personsList: usersList,
+            userIDToLookFor: FirebaseAuth.instance.currentUser!.uid,
+          );
       setState(() {
         userName = name;
       });
@@ -47,16 +52,23 @@ class _HomeAppBarState extends State<HomeAppBar> {
   void initState() {
     super.initState();
     fetchAndSetUsername();
+
+    // using addPostFrameCallback ensures it doesn't conflict with the build cycle
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      context.read<HomeAppBarProvider>().initUserData();
+    });
   }
 
   @override
   Widget build(BuildContext context) {
+    final provider = context.watch<HomeAppBarProvider>();
+
     return AnimatedContainer(
       duration: Duration(milliseconds: animationDuration),
       width: MyDimensionAdapter.getWidth(context) * 0.80,
       height: (isExpanded)
           // might be back later (still deciding)
-          ? (MyFirebaseServices.getUserType() == "admin")
+          ? (MyPersonalInfoRepository.getUserType() == "admin")
                 // ? 290
                 // : 220
                 ? 220
@@ -79,21 +91,29 @@ class _HomeAppBarState extends State<HomeAppBar> {
           Row(
             children: [
               // the user avatar/pic/icon container
-              CircleAvatar(
-                backgroundColor: Colors.blue[100],
-                child: Icon(Icons.person_rounded, color: Colors.blue),
+              GestureDetector(
+                onTap: () {
+                  showProfilePictureBottomModalSheet(context);
+                },
+                child: CircleAvatar(
+                  backgroundColor: Colors.blue[100],
+                  child: MyImageDisplayer(
+                    base64ImageString: provider.cachedImageString,
+                  ),
+                ),
               ),
               SizedBox(width: 10),
               // greeting text
               SizedBox(
                 width: MyDimensionAdapter.getWidth(context) * 0.50,
                 // this userName will be updated by setState
-                child: (userName == "User")
+                child: (userName == "...")
                     ? CircularProgressIndicator()
                     : Text(
                         // "${dotenv.env['SAMPLE_TEXT']}",
                         // "a dnbajbdjab ahbdhjabd ajbdhwbahw abdabhj",
-                        userName,
+                        // userName,
+                        provider.userName,
                         style: TextStyle(
                           // color: Colors.,
                           fontWeight: FontWeight.bold,
